# backend/src/generated/swagger.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    _36_Enums.Gender:
      type: string
      enum:
        - MALE
        - FEMALE
        - OTHER
        - PREFER_NOT_TO_SAY
    Gender:
      $ref: "#/components/schemas/_36_Enums.Gender"
    _36_Enums.StudentStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - TRANSFERRED
    StudentStatus:
      $ref: "#/components/schemas/_36_Enums.StudentStatus"
    _36_Enums.EnrollmentStatus:
      type: string
      enum:
        - TRANSFERRED
        - PRESENT
        - ABSENT
        - EXCLUDED
    EnrollmentStatus:
      $ref: "#/components/schemas/_36_Enums.EnrollmentStatus"
    EnrollmentRecordDTO:
      properties:
        id:
          type: number
          format: double
        studentId:
          type: number
          format: double
        classSectionId:
          type: number
          format: double
        schoolId:
          type: number
          format: double
          nullable: true
        withdrawnDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/EnrollmentStatus"
        isCurrent:
          type: boolean
        createdAt:
          type: string
        className:
          type: string
      required:
        - id
        - studentId
        - classSectionId
        - schoolId
        - withdrawnDate
        - status
        - isCurrent
        - createdAt
      type: object
      additionalProperties: false
    StudentResponseDTO:
      properties:
        id:
          type: number
          format: double
        schoolId:
          type: number
          format: double
        classId:
          type: number
          format: double
          nullable: true
        studentIdCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        gender:
          allOf:
            - $ref: "#/components/schemas/Gender"
          nullable: true
        status:
          $ref: "#/components/schemas/StudentStatus"
        createdAt:
          type: string
        updatedAt:
          type: string
        createdById:
          type: number
          format: double
        schoolName:
          type: string
        className:
          type: string
        currentEnrollment:
          allOf:
            - $ref: "#/components/schemas/EnrollmentRecordDTO"
          nullable: true
      required:
        - id
        - schoolId
        - classId
        - studentIdCode
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - status
        - createdAt
        - updatedAt
        - createdById
      type: object
      additionalProperties: false
    CreateStudentInput:
      properties:
        schoolId:
          type: number
          format: double
        studentIdCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        gender:
          allOf:
            - $ref: "#/components/schemas/Gender"
          nullable: true
        classId:
          type: number
          format: double
          nullable: true
        createdById:
          type: number
          format: double
      required:
        - schoolId
        - studentIdCode
        - firstName
        - lastName
      type: object
      additionalProperties: false
    PaginatedStudentResponseDTO:
      properties:
        records:
          items:
            $ref: "#/components/schemas/StudentResponseDTO"
          type: array
        total:
          type: number
          format: double
        page:
          type: number
          format: double
        limit:
          type: number
          format: double
        totalPages:
          type: number
          format: double
      required:
        - records
        - total
        - page
        - limit
        - totalPages
      type: object
      additionalProperties: false
    UpdateStudentInput:
      properties:
        schoolId:
          type: number
          format: double
        studentIdCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        gender:
          allOf:
            - $ref: "#/components/schemas/Gender"
          nullable: true
        classId:
          type: number
          format: double
          nullable: true
        status:
          $ref: "#/components/schemas/StudentStatus"
      type: object
      additionalProperties: false
    EnrollStudentInput:
      properties:
        classSectionId:
          type: number
          format: double
      required:
        - classSectionId
      type: object
      additionalProperties: false
    UnenrollStudentInput:
      properties:
        classSectionId:
          type: number
          format: double
        reason:
          type: string
      type: object
      additionalProperties: false
    TransferStudentInput:
      properties:
        fromClassSectionId:
          type: number
          format: double
        toClassSectionId:
          type: number
          format: double
        reason:
          type: string
      required:
        - toClassSectionId
      type: object
      additionalProperties: false
    _36_Enums.ClassStatus:
      type: string
      enum:
        - ACTIVE
        - ARCHIVED
    ClassStatus:
      $ref: "#/components/schemas/_36_Enums.ClassStatus"
    EnrolledStudentSummaryDTO:
      properties:
        id:
          type: number
          format: double
        studentIdCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        gender:
          allOf:
            - $ref: "#/components/schemas/Gender"
          nullable: true
        status:
          $ref: "#/components/schemas/StudentStatus"
        enrolledAt:
          type: string
      required:
        - id
        - studentIdCode
        - firstName
        - lastName
        - gender
        - status
        - enrolledAt
      type: object
      additionalProperties: false
    ClassResponseDTO:
      properties:
        id:
          type: number
          format: double
        schoolId:
          type: number
          format: double
        gradeId:
          type: number
          format: double
          nullable: true
        className:
          type: string
        section:
          type: string
        name:
          type: string
        academicYear:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ClassStatus"
        assessmentCycle:
          type: string
          nullable: true
        capacity:
          type: number
          format: double
        enrolledCount:
          type: number
          format: double
        availableSeats:
          type: number
          format: double
        createdAt:
          type: string
        updatedAt:
          type: string
        createdById:
          type: number
          format: double
        schoolName:
          type: string
        gradeName:
          type: string
        enrolledStudents:
          items:
            $ref: "#/components/schemas/EnrolledStudentSummaryDTO"
          type: array
      required:
        - id
        - schoolId
        - gradeId
        - className
        - section
        - name
        - academicYear
        - status
        - assessmentCycle
        - capacity
        - enrolledCount
        - availableSeats
        - createdAt
        - updatedAt
        - createdById
      type: object
      additionalProperties: false
    CreateClassInput:
      properties:
        schoolId:
          type: number
          format: double
        gradeId:
          type: number
          format: double
          nullable: true
        className:
          type: string
        section:
          type: string
        name:
          type: string
        academicYear:
          type: string
          nullable: true
        assessmentCycle:
          type: string
          nullable: true
        capacity:
          type: number
          format: double
        createdById:
          type: number
          format: double
      required:
        - schoolId
        - className
        - section
        - name
      type: object
      additionalProperties: false
    PaginatedClassResponseDTO:
      properties:
        records:
          items:
            $ref: "#/components/schemas/ClassResponseDTO"
          type: array
        total:
          type: number
          format: double
        page:
          type: number
          format: double
        limit:
          type: number
          format: double
        totalPages:
          type: number
          format: double
      required:
        - records
        - total
        - page
        - limit
        - totalPages
      type: object
      additionalProperties: false
    UpdateClassInput:
      properties:
        schoolId:
          type: number
          format: double
        gradeId:
          type: number
          format: double
          nullable: true
        className:
          type: string
        section:
          type: string
        name:
          type: string
        academicYear:
          type: string
          nullable: true
        assessmentCycle:
          type: string
          nullable: true
        capacity:
          type: number
          format: double
        status:
          $ref: "#/components/schemas/ClassStatus"
      type: object
      additionalProperties: false
    AssessmentStatus:
      enum:
        - Present
        - Absent
        - Partial
      type: string
    OralFlag:
      enum:
        - C0
        - C1
        - C2
        - C3
      type: string
    BaselineDomain:
      enum:
        - Vocabulary
        - Grammar
        - Phrase_Sentence
        - Listening
        - Speaking
        - Reading
        - Writing
      type: string
    SuggestedStage:
      enum:
        - S1
        - S2
        - S3
        - S4
        - S5
        - Review
        - AB
      type: string
    DomainScoreRecord:
      properties:
        id:
          type: number
          format: double
        assessmentId:
          type: number
          format: double
        domain:
          $ref: "#/components/schemas/BaselineDomain"
        item1:
          type: number
          format: double
          nullable: true
        item2:
          type: number
          format: double
          nullable: true
        item3:
          type: number
          format: double
          nullable: true
        item4:
          type: number
          format: double
          nullable: true
        item5:
          type: number
          format: double
          nullable: true
        domainScore:
          type: number
          format: double
        suggestedStage:
          $ref: "#/components/schemas/SuggestedStage"
        finalStage:
          allOf:
            - $ref: "#/components/schemas/SuggestedStage"
          nullable: true
        reviewNeeded:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - assessmentId
        - domain
        - item1
        - item2
        - item3
        - item4
        - item5
        - domainScore
        - suggestedStage
        - finalStage
        - reviewNeeded
        - createdAt
      type: object
      additionalProperties: false
    BaselineAssessmentRecord:
      properties:
        id:
          type: number
          format: double
        studentId:
          type: string
        assessmentDate:
          type: string
        assessorName:
          type: string
        status:
          $ref: "#/components/schemas/AssessmentStatus"
        keySupportFlag:
          type: string
          nullable: true
        oralFlag:
          allOf:
            - $ref: "#/components/schemas/OralFlag"
          nullable: true
        qcNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        domainScores:
          items:
            $ref: "#/components/schemas/DomainScoreRecord"
          type: array
      required:
        - id
        - studentId
        - assessmentDate
        - assessorName
        - status
        - keySupportFlag
        - oralFlag
        - qcNotes
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
...
  securitySchemes: {}
info:
  title: lumino1-backend
  version: 1.0.0
  description: Lumino1 NGO Baseline Assessment Production-Ready Express TypeScript Backend
  license:
    name: ISC
  contact:
    name: Edspectrum NGO
paths:
...
  /students:
    post:
      operationId: CreateStudent
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "400":
          description: Bad Request - Validation Error
        "409":
          description: Conflict - Student code already exists or class capacity reached
      description: Create a new Student record
      tags:
        - Students
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentInput"
    get:
      operationId: ListStudents
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStudentResponseDTO"
      description: List paginated Students with optional filters
      tags:
        - Students
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: schoolId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: classId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/StudentStatus"
        - in: query
          name: search
          required: false
          schema:
            type: string
  /students/{id}:
    get:
      operationId: GetStudentById
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "404":
          description: Student not found
      description: Get Student details by numeric ID
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: UpdateStudent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "404":
          description: Student not found
        "409":
          description: Conflict - Student code duplicate
      description: Update Student profile by ID
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentInput"
    delete:
      operationId: DeleteStudent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                required:
                  - message
                  - success
                type: object
        "404":
          description: Student not found
      description: Delete Student by ID
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
  /students/{id}/enroll:
    post:
      operationId: EnrollStudent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "400":
          description: Bad Request - Student or Class inactive
        "404":
          description: Student or Class not found
        "409":
          description: Conflict - Already enrolled or Class capacity limit reached
      description: Enroll Student into a Class Section
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollStudentInput"
  /students/{id}/unenroll:
    post:
      operationId: UnenrollStudent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "400":
          description: Bad Request - Student not currently enrolled
        "404":
          description: Student not found
      description: Unenroll Student from a Class Section
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnenrollStudentInput"
  /students/{id}/transfer:
    post:
      operationId: TransferStudent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponseDTO"
        "400":
          description: Bad Request - Invalid transfer source/target or inactive class
        "404":
          description: Student or Target Class not found
        "409":
          description: Conflict - Target class capacity limit reached or duplicate
            enrollment
      description: Transfer Student from one Class Section to another
      tags:
        - Students
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferStudentInput"
  /classes:
    post:
      operationId: CreateClass
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
        "400":
          description: Bad Request - Validation Error
        "404":
          description: School or Grade not found
        "409":
          description: Conflict - Section already exists for School/Grade
      description: Create a new Class Section
      tags:
        - Classes
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassInput"
    get:
      operationId: ListClasses
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedClassResponseDTO"
      description: List paginated Class Sections with optional filters
      tags:
        - Classes
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: schoolId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: gradeId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/ClassStatus"
        - in: query
          name: academicYear
          required: false
          schema:
            type: string
        - in: query
          name: search
          required: false
          schema:
            type: string
  /classes/{id}:
    get:
      operationId: GetClassById
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
        "404":
          description: Class Section not found
      description: Get Class Section details by numeric ID
      tags:
        - Classes
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: UpdateClass
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
        "400":
          description: Bad Request - Reduced capacity below enrolled count
        "404":
          description: Class Section not found
        "409":
          description: Conflict - Section duplicate
      description: Update Class Section by ID
      tags:
        - Classes
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassInput"
    delete:
      operationId: DeleteClass
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                required:
                  - message
                  - success
                type: object
        "404":
          description: Class Section not found
        "409":
          description: Conflict - Cannot delete class with active enrolled students
      description: Delete Class Section by ID (Enforces non-empty safeguard)
      tags:
        - Classes
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
  /classes/{id}/students:
    get:
      operationId: GetEnrolledStudents
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  totalPages:
                    type: number
                    format: double
                  page:
                    type: number
                    format: double
                  total:
                    type: number
                    format: double
                  records:
                    items:
                      $ref: "#/components/schemas/EnrolledStudentSummaryDTO"
                    type: array
                required:
                  - totalPages
                  - page
                  - total
                  - records
                type: object
        "404":
          description: Class Section not found
      description: Get paginated list of active enrolled students in a Class Section
      tags:
        - Classes
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: page
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
...
  /baseline-assessments/student/{studentId}:
    get:
      operationId: GetByStudentId
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BaselineAssessmentRecord"
                type: array
      description: Get Baseline Assessments by Student ID
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
...
servers:
  - url: /